Consolidated Preview; js_only.
