+++ assets/src/application/ui/input/InlineEditInput/InlineEditTextArea.tsx
@@ -2,10 +2,11 @@ import React from 'react';
 
 import { InlineEdit } from '@infraUI/inputs';
 import Preview from './Preview';
-import { TextAreaProps } from '../types';
+import { TextAreaProps } from './types';
 
 const InlineEditTextArea: React.FC<TextAreaProps> = ({ tooltip, ...props }) => {
-	const preview = (previewProps) => <Preview {...previewProps} tooltip={tooltip} />;
+	const preview = (previewProps) => <Preview {...previewProps} lineCount={3} tooltip={tooltip} />;
+
 	return <InlineEdit placeholder='' {...props} inputType='textarea' Preview={preview} />;
 };
