+++ assets/src/infrastructure/ui/inputs/inlineEdit/InlineEditPreview.tsx
@@ -1,15 +1,9 @@
 import React from 'react';
 import { EditablePreview as ChakraEditablePreview } from '@chakra-ui/core';
 
-import type { InlineEditPreviewProps } from './types';
+import type { PreviewProps } from './types';
 
-const InlineEditPreview: React.FC<InlineEditPreviewProps> = ({
-	isEditing,
-	onRequestEdit,
-	Preview,
-	value,
-	...props
-}) => {
+const InlineEditPreview: React.FC<PreviewProps> = ({ isEditing, onRequestEdit, Preview, value, ...props }) => {
 	return Preview ? (
 		<Preview {...props} isEditing={isEditing} onRequestEdit={onRequestEdit} value={value} />
 	) : (
